LayoutTests/fast/forms/ios/select-options-with-hr-expected.txt

This test verifies that choosing options in a <select> with an <hr> child works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS select.value is "A"
PASS select.value is "D"
PASS select.value is "E"
PASS successfullyParsed is true

TEST COMPLETE
